feat: moves to esm
