+++ package.json
@@ -11,6 +11,7 @@
     "name": "Jacob Fletcher",
     "url": "https://jacobsfletch.com"
   },
+  "type": "module",
   "bugs": {
     "url": "https://github.com/faceless-ui/scroll-info/issues",
     "email": "dev@facelessui.com"
@@ -48,9 +49,9 @@
     "@types/node": "^14.14.22",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
-    "eslint": "^8.14.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
+    "eslint": "^8.56.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jest": "^26.1.5",
     "eslint-plugin-jest-dom": "^4.0.1",
@@ -66,7 +67,7 @@
     "react-refresh": "^0.13.0",
     "react-refresh-typescript": "^2.0.4",
     "ts-loader": "^9.2.9",
-    "typescript": "^4.6.3",
+    "typescript": "^5.4.5",
     "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2",
     "webpack-dev-server": "^4.8.1"
