+++ src/ScrollInfo/index.tsx
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
-import useScrollInfo from '../useScrollInfo';
-import { IScrollInfoContext } from '../ScrollInfoProvider/context';
+import useScrollInfo from '../useScrollInfo/index.js';
+import { IScrollInfoContext } from '../ScrollInfoProvider/context.js';
 
 export type ChildFunction = (context: IScrollInfoContext) => React.ReactNode; // eslint-disable-line no-unused-vars
