+++ src/index.ts
@@ -1,5 +1,5 @@
-export { default as ScrollInfo } from './ScrollInfo';
-export { default as ScrollInfoContext } from './ScrollInfoProvider/context';
-export { default as ScrollInfoProvider } from './ScrollInfoProvider';
-export { default as useScrollInfo } from './useScrollInfo';
-export { default as withScrollInfo } from './withScrollInfo';
+export { default as ScrollInfo } from './ScrollInfo/index.js';
+export { default as ScrollInfoContext } from './ScrollInfoProvider/context.js';
+export { default as ScrollInfoProvider } from './ScrollInfoProvider/index.js';
+export { default as useScrollInfo } from './useScrollInfo/index.js';
+export { default as withScrollInfo } from './withScrollInfo/index.js';
