+++ src/types.ts
@@ -1,3 +1,3 @@
-export { ScrollInfoProviderProps } from './ScrollInfoProvider';
-export { ScrollInfoProps } from './ScrollInfo';
-export { IScrollInfoContext } from './ScrollInfoProvider/context';
+export { ScrollInfoProviderProps } from './ScrollInfoProvider/index.js';
+export { ScrollInfoProps } from './ScrollInfo/index.js';
+export { IScrollInfoContext } from './ScrollInfoProvider/context.js';
