+++ src/useScrollInfo/index.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import ScrollInfoContext, { IScrollInfoContext } from '../ScrollInfoProvider/context';
+import ScrollInfoContext, { IScrollInfoContext } from '../ScrollInfoProvider/context.js';
 
 const useScrollInfo = (): IScrollInfoContext => useContext(ScrollInfoContext);
