+++ tsconfig.json
@@ -2,8 +2,9 @@
   "compilerOptions": {
     "outDir": "./dist",
     "allowJs": true,
-    "target": "es5",
-    "module": "commonjs",
+    "target": "es6",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
     "sourceMap": true,
     "strict": true,
     "skipLibCheck": true,
